Fix CI: pin the ruff rule set instead of inheriting shifting defaults
CI had been red for three runs and I never checked it after adding the
workflow. The cause was leaving ruff's rule selection unspecified, so it
enforced whatever the current release defaults to -- rules nobody chose,
including blind-except and try-except-continue.

Those two are deliberate in the decode scripts: one corrupt audio cell must not
abort a three-hour run over hundreds of thousands of clips, and the exception
carries nothing actionable. They are now ignored for scripts/ with that reason
recorded, rather than restructuring working code to satisfy a linter.

E702 and E741 are ignored as style. Real problems were fixed rather than
silenced: a leftover unused variable in export_onnx.py from an earlier
restructuring, and three int() casts around expressions already integral.

Verified the whole workflow locally before pushing this time -- ruff, build,
install, import, CLI, twine check.

+++ scripts/decode_base.py
@@ -46,7 +46,7 @@ def decode_cell(cell):
     if w.ndim > 1:
         w = w.mean(axis=1)
     if sr != SR:
-        n = int(round(len(w) * SR / sr))
+        n = round(len(w) * SR / sr)
         w = np.interp(np.linspace(0, len(w) - 1, n), np.arange(len(w)), w).astype(np.float32)
     return w
